'grade1' and 'grade2' from int to long   (to avoid overflows) GRADING=-2 for invalid pulses   (those with too few samples to apply the shortest filter)
 SIRENA update (v.13.1.0)

+++ libsirena/scheduler.cpp
@@ -157,8 +157,8 @@ void scheduler::push_detection(TesRecord* record, int trig_reclength,
   input->event_list->energies = new double[event_list->size];
   input->event_list->avgs_4samplesDerivative = new double[event_list->size];
   input->event_list->Es_lowres = new double[event_list->size];
-  input->event_list->grades1 = new int[event_list->size];
-  input->event_list->grades2 = new int[event_list->size];
+  input->event_list->grades1 = new long[event_list->size];
+  input->event_list->grades2 = new long[event_list->size];
   input->event_list->pulse_heights = new double[event_list->size];
   //ph_ids_array
   input->event_list->grading = new int[event_list->size];
